Update on copy selected node shortcut
