+++ app/components/Icons/CopyShortcutIcon.tsx
@@ -11,8 +11,8 @@ export function CopyShortcutIcon(props: React.SVGProps<SVGSVGElement>) {
       <rect width="14" height="14" rx="1.53846" fill="currentColor" />
       <rect x="16" width="14" height="14" rx="1.53846" fill="currentColor" />
       <path
-        d="M9.59,7.86H8.72V6.14h.87A1.73,1.73,0,1,0,7.86,4.41v.87H6.14V4.41A1.73,1.73,0,1,0,4.41,6.14h.87V7.86H4.41A1.73,1.73,0,1,0,6.14,9.59h0V8.72H7.86v.87h0A1.73,1.73,0,1,0,9.59,7.86Zm0-4.31a.87.87,0,0,1,0,1.73H8.72V4.41A.87.87,0,0,1,9.59,3.55Zm-6,.86a.87.87,0,0,1,1.73,0v.87H4.41A.87.87,0,0,1,3.55,4.41Zm.86,6a.87.87,0,0,1,0-1.73h.87v.87A.87.87,0,0,1,4.41,10.45ZM6.14,7.86V6.14H7.86V7.86Zm3.45,2.59a.87.87,0,0,1-.87-.86h0V8.72h.87a.87.87,0,0,1,0,1.73Z"
-        fill="#0f172a"
+        d="M5.64,10.22H8.25V7a.39.39,0,0,1,.38-.39H10l-3-3L4,6.65H5.26A.39.39,0,0,1,5.64,7v3.18Zm3,.78H5.26a.38.38,0,0,1-.39-.39V7.43H3.11a.39.39,0,0,1-.28-.66L6.72,2.86a.39.39,0,0,1,.55,0l3.88,3.91a.38.38,0,0,1-.27.66H9v3.18a.38.38,0,0,1-.39.39Z"
+        stroke="#0f172a" stroke-width="0.35px" fill="#0f172a"
       />
       <path
         d="M23.81,9.52a1.66,1.66,0,0,0,.53-.27,1.57,1.57,0,0,0,.35-.42,1.07,1.07,0,0,0,.13-.53h1.6a2.26,2.26,0,0,1-.25,1,2.87,2.87,0,0,1-.7.85,3.35,3.35,0,0,1-1,.58,3.56,3.56,0,0,1-1.22.21,3.73,3.73,0,0,1-1.55-.31,3.2,3.2,0,0,1-1.11-.84,3.61,3.61,0,0,1-.67-1.22,4.91,4.91,0,0,1-.23-1.5V6.88a4.91,4.91,0,0,1,.23-1.5,3.54,3.54,0,0,1,.67-1.22,3.33,3.33,0,0,1,1.11-.84,3.94,3.94,0,0,1,2.83-.09,3,3,0,0,1,1,.59,2.66,2.66,0,0,1,.67.92,2.94,2.94,0,0,1,.23,1.17h-1.6a1.48,1.48,0,0,0-.45-1.07,1.65,1.65,0,0,0-.52-.33,1.75,1.75,0,0,0-.65-.12,1.59,1.59,0,0,0-1.44.78,2.27,2.27,0,0,0-.31.8,4.53,4.53,0,0,0-.09.91v.24a4.63,4.63,0,0,0,.09.92,2.46,2.46,0,0,0,.3.8,1.67,1.67,0,0,0,.56.56,1.63,1.63,0,0,0,.89.22A2.05,2.05,0,0,0,23.81,9.52Z"
